[FE] refactor: Redux 디렉토리 구조 변경
- issue (#134)
