+++ client/src/components/Header/RoleSwitch/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import styled from '@emotion/styled';
 import { userHttpClient } from '@api';
 import { useAppSelector, useAppDispatch } from '@hooks';
-import { selectUserRole, changeUserRole } from '@store/slices/userSlice';
+import { selectUserRole, changeUserRole } from '@store/user/globalSlice';
 
 function RoleSwitch(): JSX.Element {
   const dispatch = useAppDispatch();
