+++ client/src/components/Layout/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Header, Loader } from '@components';
 import Router from '@core/Router';
 import { useAppSelector } from '@hooks';
-import { selectLoadingState } from '@store/slices/utilSlice';
+import { selectLoadingState } from '@store/util/Slice';
 
 const DefaultLayout = (): JSX.Element => {
   const isLoading = useAppSelector(selectLoadingState);
