+++ client/src/pages/ProfilePage/profileBox/ProfileMentorStackBox.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import styled from '@emotion/styled';
 import ProfileContainer from './ProfileBoxContainer';
 import { useAppSelector } from '@hooks';
-import { selectProfileData } from '@store/slices/profileDataSlice';
-import { selectUser } from '@store/slices/userSlice';
+import { selectProfileData } from '@store/user/profileSlice';
+import { selectUser } from '@store/user/globalSlice';
 
 interface Props {
   showRequestModal: () => void;
@@ -20,58 +20,50 @@ function ProfileMentorStackBox({ showCreateModal, showRequestModal }: Props): JS
   const profile = useAppSelector(selectProfileData);
   const user = useAppSelector(selectUser);
 
-  const getContentsElement = (type: number) =>{
-    switch(type){
+  const getContentsElement = (type: number) => {
+    switch (type) {
       case ViewType.OTHER:
         return (
           <ProfileContainer title="멘토링스택">
-          <TechStackContainer>
-            {profile.mentoringStack.map((techStackName, idx) => (
-              <TechStackItem key={idx}>{techStackName}</TechStackItem>
-            ))}
-          </TechStackContainer>
-        </ProfileContainer>
-        )
+            <TechStackContainer>
+              {profile.mentoringStack.map((techStackName, idx) => (
+                <TechStackItem key={idx}>{techStackName}</TechStackItem>
+              ))}
+            </TechStackContainer>
+          </ProfileContainer>
+        );
       case ViewType.ME_MENTOR:
         return (
           <ProfileContainer title="멘토링스택">
-          <MentoringRequestButton onClick={showRequestModal}>
-            멘토요청 리스트
-          </MentoringRequestButton>
-          <TechStackContainer>
-            {profile.mentoringStack.map((techStackName, idx) => (
-              <TechStackItem key={idx}>{techStackName}</TechStackItem>
-            ))}
-          </TechStackContainer>
-        </ProfileContainer>
-        )
+            <MentoringRequestButton onClick={showRequestModal}>
+              멘토요청 리스트
+            </MentoringRequestButton>
+            <TechStackContainer>
+              {profile.mentoringStack.map((techStackName, idx) => (
+                <TechStackItem key={idx}>{techStackName}</TechStackItem>
+              ))}
+            </TechStackContainer>
+          </ProfileContainer>
+        );
       case ViewType.ME_NOT_MENTOR:
         return (
           <ProfileContainer title="">
-          <MentorRegisterTitle>멘토가 되어주세요!</MentorRegisterTitle>
-          <MentorRegisterDesc>
-            간단한 기술스택을 등록을 통해 멘토가 될 수 있어요!
-          </MentorRegisterDesc>
-          <MentorRegisterButton onClick={showCreateModal}>멘토 신청하기</MentorRegisterButton>
-        </ProfileContainer>
-        )
+            <MentorRegisterTitle>멘토가 되어주세요!</MentorRegisterTitle>
+            <MentorRegisterDesc>
+              간단한 기술스택을 등록을 통해 멘토가 될 수 있어요!
+            </MentorRegisterDesc>
+            <MentorRegisterButton onClick={showCreateModal}>멘토 신청하기</MentorRegisterButton>
+          </ProfileContainer>
+        );
     }
+  };
+  const getViewType = (): number => {
+    if (profile.userId !== user.id) return ViewType.OTHER;
+    else if (profile.isMentor) return ViewType.ME_MENTOR;
+    else return ViewType.ME_NOT_MENTOR;
+  };
 
-  }
-  const getViewType = ():number =>{
-    if(profile.userId !== user.id)
-      return ViewType.OTHER;
-    else if(profile.isMentor)
-      return ViewType.ME_MENTOR;
-    else
-      return ViewType.ME_NOT_MENTOR;
-  }
-
-  return (
-    <>
-      {getContentsElement(getViewType())}
-    </>
-  );
+  return <>{getContentsElement(getViewType())}</>;
 }
 
 const MentorRegisterTitle = styled.h3`
