+++ client/src/store/user/globalSlice.ts
