+++ client/src/store/util/Slice.ts
